refactor: extract LineEditBuffer from LineEditor
